refactoring: 함수 추출 및 ref id로 변경
