+++ src/app/page.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import { GSAPConfig } from '@/lib/gsap/GSAPConfig';
-import useImagesLoading from '@/hooks/useImagesLoading';
 import NavBar from '@/feature/nav/NavBar';
 import AboutLayout from '@/feature/about/layouts/AboutLayout';
 import SkillLayout from '@/feature/skill/layouts/SkillLayout';
@@ -9,11 +7,7 @@ import ProjectLayout from '@/feature/project/layouts/ProjectLayout';
 import CareerLayout from '@/feature/career/layouts/CareerLayout';
 import ContactLayout from '@/feature/contact/layouts/ContactLayout';
 
-GSAPConfig.initialize();
-
 export default function Home() {
-  const { loadingCount } = useImagesLoading();
-
   return (
     <div>
       <main
