+++ src/feature/contact/layouts/ContactLayout.module.scss
@@ -1,8 +1,15 @@
 @import "@/styles/variables.scss";
 
 .contactLayoutWrapper {
+  display: flex;
+
+  justify-content: center;
+  align-items: center;
+
   position: sticky;
   bottom: 0;
+  height: 80vh;
+  
   background-color: $color-primary;
 }
